Move _tx_neutral_date to helpers
