+++ AIPscan/Aggregator/task_helpers.py
@@ -3,7 +3,26 @@
 """Collects a number of reusable components of tasks.py. Also ensures
 the module remains clean and easy to refactor over time.
 """
+from datetime import datetime
 import os
+
+from dateutil.parser import parse, ParserError
+
+
+def _tz_neutral_date(date):
+    """Convert inconsistent dates consistently. Dates are round-tripped
+    back to a Python datetime object as anticipated by the database.
+    Where a date is unknown or can't be parsed, we return the UNIX EPOCH
+    in lieu of another sensible value.
+    """
+    EPOCH = datetime.strptime("1970-01-01T00:00:01", "%Y-%m-%dT%H:%M:%S")
+    try:
+        date = parse(date)
+        date = date.strftime("%Y-%m-%dT%H:%M:%S")
+        date = datetime.strptime(date, "%Y-%m-%dT%H:%M:%S")
+    except ParserError:
+        date = EPOCH
+    return date
 
 
 def get_mets_url(api_url, package_uuid, path_to_mets):
