+++ AIPscan/Aggregator/tasks.py
@@ -26,25 +26,11 @@ from AIPscan.models import (
     get_mets_tasks,
 )
 
-from AIPscan.Aggregator.task_helpers import get_mets_url, create_numbered_subdirs
-
-from dateutil.parser import parse, ParserError
-
-
-def _tz_neutral_date(date):
-    """Convert inconsistent dates consistently. Dates are round-tripped
-    back to a Python datetime object as anticipated by the database.
-    Where a date is unknown or can't be parsed, we return the UNIX EPOCH
-    in lieu of another sensible value.
-    """
-    EPOCH = datetime.strptime("1970-01-01T00:00:01", "%Y-%m-%dT%H:%M:%S")
-    try:
-        date = parse(date)
-        date = date.strftime("%Y-%m-%dT%H:%M:%S")
-        date = datetime.strptime(date, "%Y-%m-%dT%H:%M:%S")
-    except ParserError:
-        date = EPOCH
-    return date
+from AIPscan.Aggregator.task_helpers import (
+    get_mets_url,
+    create_numbered_subdirs,
+    _tz_neutral_date,
+)
 
 
 def write_packages_json(count, timestampStr, packages):
